meta-scm-npcm845: increase adc sensor max reading
To avoid the sensor reading and threshold error after IPMI scaling.
The current ADC max reading is const value 1.8 / [scale factor].
When we read the 1V8 power, we will set scale factor = 1, and this time
max reading will be 1.8. But in real world the power voltage may greater
than 1.8, may 1.85 or 1.9. In this case, we will recieve error reading
after IPMI scaling.

Signed-off-by: Brian_Ma <chma0@nuvoton.com>