meta-scm-npcm845: WA: fix bmcweb SOL function fail
Due to obmc console is moving configurations to D-Bus to support
multi-console. But its functions is not totally ready and the bmcweb side
change is reviewing. Update the bmcweb socket name as the obmc-console
default value to fix the SOL feature cannot work issue on bmcweb.

Note. This just is workaround patch. After the changes for bmcweb is
merged, we should remove this commit.

Ref:
obmc-console: Provide a default value for `console-id`
https://gerrit.openbmc.org/c/openbmc/obmc-console/+/63179

Use console information from DBUS
https://gerrit.openbmc.org/c/openbmc/bmcweb/+/62525

Signed-off-by: Brian Ma <chma0@nuvoton.com>